+++ app/src/main/java/com/andaagii/tacomamusicplayer/screen/PermissionDeniedScreen.kt
@@ -1,4 +1,4 @@
-package com.andaagii.tacomamusicplayer.composables
+package com.andaagii.tacomamusicplayer.screen
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
