+++ app/src/main/java/com/andaagii/tacomamusicplayer/state/AlbumTabState.kt
@@ -8,7 +8,7 @@ import com.andaagii.tacomamusicplayer.util.SortingUtil
  * Single [StateFlow] payload for the albums browsing page.
  *
  * Bundling [albums], [sorting], and [layout] into one object ensures that
- * [com.andaagii.tacomamusicplayer.composables.AlbumListScreen] reacts to all three
+ * [com.andaagii.tacomamusicplayer.screen.AlbumListScreen] reacts to all three
  * pieces of state atomically — a sort change and a layout change emitted together produce
  * exactly one UI update rather than two.
  *
